Fix several issue
Use $xoopsDB to get a database connection, so we will work in legacy mode
in 2.6 for now.

Change useTable to return boolean true if table exists, and false if it does not.
Will no longer add the table if it doesn't exist, so we can use it to manage
changes to existing tables.

Change addColumn and alterColumn argument order to make $position optional